requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==4.2.4
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.7
python-dotenv==0.19.2
pytz==2023.3.post1
sqlparse==0.4.2
whitenoise==5.3.0
